update clientchainId type from uint16 to uint32

+++ src/core/ExocoreGateway.sol
@@ -115,7 +115,7 @@ contract ExocoreGateway is
         (bool success, bytes memory responseOrReason) = DEPOSIT_PRECOMPILE_ADDRESS.call(
             abi.encodeWithSelector(
                 DEPOSIT_FUNCTION_SELECTOR,
-                uint16(srcChainId), // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
+                srcChainId, // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
                 token,
                 depositor,
                 amount
@@ -142,7 +142,7 @@ contract ExocoreGateway is
         (bool success, bytes memory responseOrReason) = WITHDRAW_PRINCIPLE_PRECOMPILE_ADDRESS.call(
             abi.encodeWithSelector(
                 WITHDRAW_PRINCIPLE_FUNCTION_SELECTOR,
-                uint16(srcChainId), // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
+                srcChainId, // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
                 token,
                 withdrawer,
                 amount
@@ -169,7 +169,7 @@ contract ExocoreGateway is
         (bool success, bytes memory responseOrReason) = CLAIM_REWARD_PRECOMPILE_ADDRESS.call(
             abi.encodeWithSelector(
                 CLAIM_REWARD_FUNCTION_SELECTOR,
-                uint16(srcChainId), // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
+                srcChainId, // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
                 token,
                 withdrawer,
                 amount
@@ -192,7 +192,7 @@ contract ExocoreGateway is
         (bool success,) = DELEGATION_PRECOMPILE_ADDRESS.call(
             abi.encodeWithSelector(
                 DELEGATE_TO_THROUGH_CLIENT_CHAIN_FUNCTION_SELECTOR,
-                uint16(srcChainId), // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
+                srcChainId, // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
                 lzNonce,
                 token,
                 delegator,
@@ -215,7 +215,7 @@ contract ExocoreGateway is
         (bool success,) = DELEGATION_PRECOMPILE_ADDRESS.call(
             abi.encodeWithSelector(
                 UNDELEGATE_FROM_THROUGH_CLIENT_CHAIN_FUNCTION_SELECTOR,
-                uint16(srcChainId), // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
+                srcChainId, // TODO: Casting srcChainId from uint32 to uint16 should be fixed after exocore network fix source chain id type
                 lzNonce,
                 token,
                 delegator,
